# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Sport {
  key: string;
  title: string;
  group: string;
}

export type BetStatus = typeof BetStatus[keyof typeof BetStatus];


export const BetStatus = {
  pending: 'pending',
  won: 'won',
  lost: 'lost',
  push: 'push',
} as const;

export type BetMarket = typeof BetMarket[keyof typeof BetMarket];


export const BetMarket = {
  h2h: 'h2h',
  spreads: 'spreads',
  totals: 'totals',
} as const;

export interface EdgeOpportunity {
  gameId: string;
  sport: string;
  commenceTime: string;
  homeTeam: string;
  awayTeam: string;
  market: BetMarket;
  selection: string;
  /** @nullable */
  point: number | null;
  americanOdds: number;
  book: string;
  fairOdds: number;
  evPercent: number;
}

export interface Bet {
  id: number;
  sport: string;
  gameId: string;
  commenceTime: string;
  homeTeam: string;
  awayTeam: string;
  market: BetMarket;
  selection: string;
  /** @nullable */
  point: number | null;
  americanOdds: number;
  units: number;
  /** @nullable */
  fairOdds: number | null;
  /** @nullable */
  evPercent: number | null;
  /** @nullable */
  book: string | null;
  /** @nullable */
  closingOdds: number | null;
  /** @nullable */
  clvPercent: number | null;
  status: BetStatus;
  /** @nullable */
  pnl: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface BetInput {
  sport: string;
  gameId: string;
  commenceTime: string;
  homeTeam: string;
  awayTeam: string;
  market: BetMarket;
  selection: string;
  /** @nullable */
  point: number | null;
  americanOdds: number;
  /** @minimum 0 */
  units: number;
  /** @nullable */
  fairOdds?: number | null;
  /** @nullable */
  evPercent?: number | null;
  /** @nullable */
  book?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface BetUpdate {
  status?: BetStatus;
  americanOdds?: number;
  /** @minimum 0 */
  units?: number;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface SportBreakdown {
  sport: string;
  bets: number;
  won: number;
  lost: number;
  push: number;
  pending: number;
  roiPercent: number;
  pnl: number;
}

export interface DashboardSummary {
  totalBets: number;
  won: number;
  lost: number;
  push: number;
  pending: number;
  totalUnits: number;
  totalPnl: number;
  roiPercent: number;
  /** @nullable */
  avgClvPercent: number | null;
  clvSampleSize: number;
  bySport: SportBreakdown[];
}

export type ListEdgesParams = {
sport: string;
minEdgePercent?: number;
};

export type ListBetsParams = {
status?: BetStatus;
};