# lib/api-zod/src/generated/types/bet.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BetMarket } from './betMarket';
import type { BetStatus } from './betStatus';

export interface Bet {
  id: number;
  sport: string;
  gameId: string;
  commenceTime: Date;
  homeTeam: string;
  awayTeam: string;
  market: BetMarket;
  selection: string;
  /** @nullable */
  point: number | null;
  americanOdds: number;
  units: number;
  /** @nullable */
  fairOdds: number | null;
  /** @nullable */
  evPercent: number | null;
  /** @nullable */
  book: string | null;
  /** @nullable */
  closingOdds: number | null;
  /** @nullable */
  clvPercent: number | null;
  status: BetStatus;
  /** @nullable */
  pnl: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
}