# lib/api-zod/src/generated/types/betMarket.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BetMarket = typeof BetMarket[keyof typeof BetMarket];


export const BetMarket = {
  h2h: 'h2h',
  spreads: 'spreads',
  totals: 'totals',
} as const;