# lib/api-zod/src/generated/types/betUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BetStatus } from './betStatus';

export interface BetUpdate {
  status?: BetStatus;
  americanOdds?: number;
  /** @minimum 0 */
  units?: number;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  notes?: string | null;
}