# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SportBreakdown } from './sportBreakdown';

export interface DashboardSummary {
  totalBets: number;
  won: number;
  lost: number;
  push: number;
  pending: number;
  totalUnits: number;
  totalPnl: number;
  roiPercent: number;
  /** @nullable */
  avgClvPercent: number | null;
  clvSampleSize: number;
  bySport: SportBreakdown[];
}