# lib/api-zod/src/generated/types/edgeOpportunity.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BetMarket } from './betMarket';

export interface EdgeOpportunity {
  gameId: string;
  sport: string;
  commenceTime: Date;
  homeTeam: string;
  awayTeam: string;
  market: BetMarket;
  selection: string;
  /** @nullable */
  point: number | null;
  americanOdds: number;
  book: string;
  fairOdds: number;
  evPercent: number;
}